Add support for silos with alternate recipes
Some planets add alternate rocket part recipes,
with some being required. This patch will read the recipe
(and quality) of the chosen rocket silo when determining
what to consume when supplying a space platform.

This does not (yet) update the GUI to show the available
rocket parts based on the current recipe.