Avoid fulfilling requests with incoming pods
If using Quantum Fabricator alongside a traditional logistics
network, it was possible for both to attempt to fulfill the
same requests, resulting in more items being sent than was
desired, especially for items with a large rocket capacity.

In addition, when checking the hub inventory, it only
checked the normal quality version of an item rather than
the quality requested, so items that were delivered weren't
counted resulting in Quantum Fabricator to attempt to fulfill
the request again, potentially overflowing the space
platform the extra items.